pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "portsense"
version = "0.1.0"
description = "Port scan analysis tool scaffold (parsing local files, assessing risk, and reporting)."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Portsense Authors" }
]
dependencies = [
  "typer>=0.9.0",
  "pydantic>=2.6.0",
  "PyYAML>=6.0.0",
]

[project.scripts]
portsense = "portsense.cli:main"

[tool.setuptools]
packages = { find = { where = ["."], include = ["portsense*"] } }

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["."]
